refactor: Dropdown에서 추천 검색어 리스트 props 제거
