+++ src/@types/dropdown.ts
@@ -1,14 +1,3 @@
-import { Suggestion } from './search';
-
-export interface DropdownProps {
-  isOpen: boolean;
-  keyword: string;
-  setKeyword: (keyword: string) => void;
-  activeIndex: number;
-  setActiveIndex: (index: number) => void;
-  suggestions: Suggestion[];
-}
-
 export interface DropdownItemProps {
   keyword: string;
   isActive: boolean;
